refactor: centralize name parser package
