--- packages/frontend/src/name-parser/character-names.js
@@ -1,241 +1,0 @@
-module.exports =[
-//kankore  http://dunkel.halfmoon.jp/kancolle/
-"長門",
-"陸奥",
-"伊勢",
-"日向",
-"雪風",
-"赤城",
-"加賀",
-"蒼龍",
-"飛龍",
-"島風",
-"吹雪",
-"白雪",
-"初雪",
-"深雪",
-"叢雲",
-"磯波",
-"綾波",
-"敷波",
-"大井",
-"北上",
-"金剛",
-"比叡",
-"榛名",
-"霧島",
-"鳳翔",
-"扶桑",
-"山城",
-"天龍",
-"龍田",
-"龍驤",
-"睦月",
-"如月",
-"皐月",
-"文月",
-"長月",
-"菊月",
-"三日月",
-"望月",
-"球磨",
-"多摩",
-"木曾",
-"長良",
-"五十鈴",
-"名取",
-"由良",
-"川内",
-"神通",
-"那珂",
-"千歳",
-"千代田",
-"最上",
-"古鷹",
-"加古",
-"青葉",
-"妙高",
-"那智",
-"足柄",
-"羽黒",
-"高雄",
-"愛宕",
-"摩耶",
-"鳥海",
-"利根",
-"筑摩",
-"飛鷹",
-"隼鷹",
-"朧",
-"曙",
-"漣",
-"潮",
-"暁",
-"響",
-"雷",
-"電",
-"初春",
-"子日",
-"若葉",
-"初霜",
-"白露",
-"時雨",
-"村雨",
-"夕立",
-"五月雨",
-"涼風",
-"朝潮",
-"大潮",
-"満潮",
-"荒潮",
-"霰",
-"霞",
-"陽炎",
-"不知火",
-"黒潮",
-"祥鳳",
-"千歳甲",
-"千代田甲",
-"千歳航",
-"千代田航",
-"翔鶴",
-"瑞鶴",
-"鬼怒",
-"阿武隈",
-"夕張",
-"瑞鳳",
-"三隈",
-"初風",
-"舞風",
-"衣笠",
-"伊19",
-"鈴谷",
-"熊野",
-"伊168",
-"伊58",
-"伊8",
-"大和",
-"秋雲",
-"夕雲",
-"巻雲",
-"長波",
-"阿賀野",
-"能代",
-"矢矧",
-"酒匂",
-"武蔵",
-"Верный",
-"大鳳",
-"香取",
-"伊401",
-"あきつ丸",
-"神威",
-"まるゆ",
-"弥生",
-"卯月",
-"磯風",
-"浦風",
-"谷風",
-"浜風",
-"Bismarck",
-"Bismarck zwei",
-"Z1",
-"Z3",
-"Prinz Eugen",
-"Bismarck drei",
-"Z1 zwei",
-"Z3 zwei",
-"天津風",
-"明石",
-"大淀",
-"大鯨",
-"龍鳳",
-"時津風",
-"雲龍",
-"天城",
-"葛城",
-"春雨",
-"早霜",
-"清霜",
-"朝雲",
-"山雲",
-"野分",
-"秋月",
-"照月",
-"初月",
-"高波",
-"朝霜",
-"U-511",
-"Graf Zeppelin",
-"Saratoga",
-"呂500",
-"Warspite",
-"Iowa",
-"Littorio",
-"Roma",
-"Libeccio",
-"Aquila",
-"秋津洲",
-"Italia",
-"Zara",
-"Pola",
-"瑞穂",
-"沖波",
-"風雲",
-"嵐",
-"萩風",
-"親潮",
-"山風",
-"海風",
-"江風",
-"速吸",
-"鹿島",
-"神風",
-"朝風",
-"春風",
-"松風",
-"旗風",
-"天霧",
-"狭霧",
-"水無月",
-"伊26",
-"浜波",
-"藤波",
-"浦波",
-"Commandant Teste",
-"Richelieu",
-"伊400",
-"伊13",
-"伊14",
-"Zara due",
-"Гангут",
-"Ташкент",
-"Ark Royal",
-"Гангут два",
-"占守",
-"国後",
-"Jervis",
-"春日丸",
-"神鷹",
-"Luigi Torelli",
-"大鷹",
-"岸波",
-"UIT-25",
-"伊504",
-"涼月",
-"択捉",
-"松輪",
-"佐渡",
-"対馬",
-"日振",
-"大東",
-"福江",
-"Nelson",
-"Gotland",
-"Maestrale",
-
-"穂乃果",
-"ことり",
-
-"エルフ"
-].filter(e => e.length > 1);
