--- packages/frontend/src/name-parser/name-parser-config.js
@@ -1,63 +1,0 @@
-module.exports.same_tag_regs_table = {
-    "東方Project": [/^東方$/, /Touhou\s*Project/, /東方project/],
-    "オリジナル": [/^Original$/],
-    "Kanon": [/カノン|Kanon/],
-    "艦これ": [/艦これ|舰これ/, /艦隊これくしょん/, /Kantai\s*Collection/, /KanColle/],
-    "ラブライブ!": [/Love Live/, /ラブライブ/],
-    "ラブライブ!サンシャイン!!": [/ラブライブ.*サンシャイン.*/],
-    "プリンセスコネクト!Re:Dive": [/プリンセスコネクト.*Re.*Dive/],
-    "Fate／Grand Order": [/Fate.*Grand.*Order/, /FGO/],
-    "Fate／Stay Night": [/Fate.*Stay.*Night/],
-    "Fate／Zero": [/Fate.*Zero/],
-    "Fate／kaleid liner プリズマ☆イリヤ": [/Fate.*kaleid.*liner.*プリズマ.*イリヤ/, /Fate.*kaleid.*liner/, /プリズマ.*イリヤ/],
-    "Fate": [/^Fate\s*／*\w+/],
-    "アイドルマスター": [/アイドルマスタ/, /DOL.*M@STER/, /dol.*master/, /アイマス/],
-    "アイドルマスター シンデレラガールズ": [/アイドルマスター.*シンデレラガールズ/, /DOLM@STER.*CINDERELLA.*GIRLS/],
-    "アイドルマスター ミリオンライブ": [/アイドルマスター.*ミリオン/, /ミリオンライブ/],
-    "アイドルマスター シャイニーカラーズ": [/アイドルマスター.*シャイニーカラーズ/],
-    "アズールレーン": [/Azur Lane/],
-    "ガールズ&パンツァー": [/Girls.*nd.*Panzer/],
-    "けいおん": [/けいおん/, /K-ON/],
-    "プリキュア": [/プリキュア/],
-    "To LOVEる": [/To.*LOVEる/, /To.*LOVE.*ru/],
-    "魔法少女まどか☆マギカ": [/まどか.*マギカ|PuellaMagiMadoka/],
-    "アイカツ!": [/アイカツ.*/],
-    "エヴァンゲリオン": [/エヴァンゲリオン/, /^エヴァ$/, /Evangelion/],
-    "Angel Beats": [/Angel.*Beats.*/, /エンジェル.*ビーツ/],
-    "Dead Or Alive": [/Dead Or Alive/, /デッド.*オア.*アライヴ/, /DEADorALIVE/],
-    "IS ＜インフィニット・ストラトス＞": [/S.*インフィニット.*ストラトス.*/, /インフィニット.*ストラトス/],
-    "D.C～ダ・カーポ": [/D\.C\./, /ダ.*カーポ/],
-    "Dog Days": [/Dog.*Days.*/],
-    "Dream C Club": [/Dream.*C.*Club/],
-    "ガンダム ": [/ガンダム|gundam/],
-    "コードギアス": [/コードギアス/],
-    "ご注文はうさぎですか": [/ご注文はうさぎですか/],
-    "ソードアート・オンライン": [/ソードアート.*オンライン/, /Sword\s*Art\s*Online/],
-    "ダンガンロンパ": [/ダンガンロンパ/],
-    "ドラゴンクエスト": [/ドラゴンクエスト/, /Dragon\s*Quest/],
-    "ファイナルファンタジー": [/ファイナルファンタジー|FinalFantasy/],
-    "咲 -Saki-": [/咲.*Saki.*/],
-    "咲-Saki- 阿知賀編": [/咲.*Saki.*阿知賀編/],
-    "聖剣伝説": [/聖剣伝説/],
-    "閃乱カグラ": [/閃乱カグラ/],
-    "魔法少女リリカルなのは": [/魔法少女リリカルなのは.*|Nanoha$|^なのは$/],
-    "キング・オブ・ファイター": [/キング\.オブ\.ファイター/, /^KOF$/],
-    "ファイアーエムブレム": [/ファイアーエムブレム/],
-    "ファンタシースター": [/ファンタシースター|PhantasyStar/],
-    "To Heart 2": [/To.*Heart.*2|トゥハート2/]
-}
-
-module.exports.not_author_but_tag = [
-    "同人音声",
-    "同人誌",
-    "アンソロジー",
-    "DL版",
-    "よろず",
-    "成年コミック",
-    "Pixiv",
-    "アーティスト",
-    "雑誌",
-    "English",
-    "Chinese",
-    "320K"
-]
