--- packages/frontend/src/name-parser/readme.txt
@@ -1,4 +1,0 @@
-all in one是给tamper monkey require用的
-
-
-npx webpack --config webpack.config.js
