+++ packages/name-parser/ParserUtil.js
