+++ packages/name-parser/all_in_one/index.js
