+++ packages/name-parser/index.js
