+++ packages/name-parser/name-parser-config.js
