+++ packages/name-parser/webpack.config.js
